backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
supabase==2.5.1
python-dotenv==1.0.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pytest==8.2.2
gunicorn==22.0.0
